drivers: video: Add pwdn pin support on ov2640
The current ov2640 driver does not support pwdn pin support, so if there
is a separate pin to control pwdn, there will be problems. It is
necessary to add zephyr, deferreded-init tags in the device tree to
manually initialize, which is inconvenient. Tested on dnesp32s3 from
alientek, it works